Extract editor blocks inline model
