+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.98",
-  "tag": "v3.4.98",
+  "version": "3.4.99",
+  "tag": "v3.4.99",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.97 <3.4.98"
+      ">=3.4.98 <3.4.99"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.97 first."
+    "message": "Update to v3.4.98 first."
   }
 }
