+++ scripts/test-system-release-package.sh
@@ -428,6 +428,11 @@ if ! grep -qx "press-system-${version}/assets/js/editor-blocks-model.js" "${entr
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/editor-blocks-inline-model.js" "${entries_file}"; then
+  echo "expected package to include editor blocks inline model boundary code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-blocks-runtime.js" "${entries_file}"; then
   echo "expected package to include editor blocks runtime boundary code" >&2
   exit 1
